Fix: sorting exception handling
